% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Prelim.R
\name{exact.ci}
\alias{exact.ci}
\title{Exact Confidence Intervals}
\usage{
exact.ci(p.adjust, size.x, size.y, i)
}
\arguments{
\item{p.adjust}{A vector of concordance indices, these are proportions for which you want to calculate the confidence interval.}

\item{size.x}{A vector of the number of series used to calculate each index in matrix X.}

\item{size.y}{A vector of the number of series used to calculate each index in matrix Y.}

\item{i}{A recursive variables (i.e. use this function in Apply.)}
}
\value{
A vector with 2 elements indicating the lower and upper confidence intervals for the concordance indices.
}
\description{
Calculate the exact confidence intervals for proportions, based in Pearson's approach.
}

