% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/series_bootstrapped.R
\docType{data}
\name{names.stat}
\alias{names.stat}
\title{names.stat}
\format{\preformatted{ chr [1:2] "mean" "median"
}}
\usage{
names.stat
}
\description{
names.stat
}
\keyword{datasets}

