\name{precipitation}
\alias{precipitation}
\docType{data}
\title{
Monthly Precipitation Data
}
\description{
Monthly Precipitation Data
}
\usage{data(precipitation)}
\format{
  A data frame with 153 observations on the following 13 variables.
  \describe{
    \item{\code{YEAR}}{a numeric vector}
    \item{\code{Jan}}{a numeric vector}
    \item{\code{Feb}}{a numeric vector}
    \item{\code{Mar}}{a numeric vector}
    \item{\code{Apr}}{a numeric vector}
    \item{\code{May}}{a numeric vector}
    \item{\code{Jun}}{a numeric vector}
    \item{\code{Jul}}{a numeric vector}
    \item{\code{Aug}}{a numeric vector}
    \item{\code{Sep}}{a numeric vector}
    \item{\code{Oct}}{a numeric vector}
    \item{\code{Nov}}{a numeric vector}
    \item{\code{Dec}}{a numeric vector}
  }
}
\source{
NIWA (The National Institute of Water and Atmosphere Research, New Zealand)
}
\examples{
data(precipitation)
## maybe str(precipitation) ; plot(precipitation) ...
}
\keyword{datasets}
