% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/site_chron.R
\name{site.chron}
\alias{site.chron}
\title{site.chron}
\usage{
site.chron(data.m, aver.by.tree.input)
}
\arguments{
\item{data.m}{A data matrix contains standardized ring width indices}

\item{aver.by.tree.input}{If True then averages tree series, then average tree means. If False then average all series}
}
\value{
\item{aver.site}{The average for the site.}
\item{std.site}{The standard deviation of the site.}
\item{aver.tree}{A time series for average of each tree.}
\item{number.trees}{The total number of trees used to produce the site chronology.}
\item{core.per.tree}{The number of cores for each tree.}
}
\description{
This function calculate the site chronology for the data matrix
}
\examples{
\dontrun{
site.full <- site.chron(spline200.sub0.2000.n$sub.series.stand, aver.by.tree=FALSE)
}
}

