\name{SOI}
\alias{SOI}
\docType{data}
\title{SOI data}
\description{
Southern Oscillation Index Data 
}
\usage{data(SOI)}
\format{
  The format is:
 chr "SOI"
}
\source{
BOM (Australian Bureau of Meteorology)
}
\examples{
data(SOI)
## maybe str(SOI) ; plot(SOI) ...
}
\keyword{datasets}
