% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/series_bootstrapped.R
\name{stat}
\alias{stat}
\title{Stat}
\usage{
stat(data)
}
\arguments{
\item{data}{data matrix (produced within series.bootstrapped)}
}
\value{
results required output for series.bootstrapped
}
\description{
To be used internally in series.bootstrapped this is the summary that is calculated during bootstrapping.
}

