% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Prelim.R
\name{SumNotNa}
\alias{SumNotNa}
\title{Sum non Na's}
\usage{
SumNotNa(x)
}
\arguments{
\item{x}{Vector to be counted containing Na's}
}
\value{
The sum of the non na's values.
}
\description{
Count the number of non na values. This is an internal function used in an apply function.
}

